### 2025-06-24 #8471 @kyleniemeyer
Hello @jake-w-liu, due to the relatively small size and wrapper nature of your submission, the JOSS editorial board is going to take a closer look to make sure it meets our requirements.

In the meantime, can you help with explaining how this differentiates from PlotlyJS itself? Normally, wrappers are not in scope for JOSS; I'm also wondering how this is [research software based on our definition](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software):
> solves complex modeling problems in a scientific context (physics, mathematics, biology, medicine, social science, neuroscience, engineering); supports the functioning of research instruments or the execution of research experiments; extracts knowledge from large data sets; offers a mathematical library, or similar.

(This isn't a statement about the software being useful or well-written.)
