### 2025-08-01 #8471 @jake-w-liu
Hello @kyleniemeyer,
I noticed that there is another journal titled Journal of Open Source Education (JOSE).
I’m wondering if the topic of this paper might be more appropriate for that venue.
If so, I can withdraw the current submission and submit there.
Thank you!
