### 2025-08-12 #8471 @kyleniemeyer
@jake-w-liu sorry for the delay in responding - the editorial board voted in favor of your submission as in scope for JOSS, so we can proceed to review here. The only issue right now is a lack of editorial availability to handle this, so it will be on our waitlist until someone frees up.

Of course, if you'd like to withdraw and submit to JOSE, you can do so, but we will review your submission.
