### 2025-08-13 4622733
remove grid from heatmap
