### 2025-08-13 5788212
add equalar in heatmap
