+++ Project.toml
@@ -5,7 +5,6 @@ version = "1.0.1"
 
 [deps]
 BatchAssign = "e837618a-1c15-48db-beb0-68297dffdb28"
-FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 Infiltrator = "5903a43b-9cc3-4c30-8d17-598619ec4e9b"
 MeshGrid = "ebf956a0-ef5e-43be-9fb1-27952996e635"
 PlotlyJS = "f0f68f2c-4968-5e81-91da-67840de0976a"
@@ -13,10 +12,9 @@ Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 
 [compat]
 BatchAssign = "1.0.0"
-FFTW = "1.8.0"
 Infiltrator = "1.9.1"
 MeshGrid = "1.0.3"
-PlotlyJS = "0.18.15"
+PlotlyJS = "0.18.17"
 Reexport = "1.2.2"
 julia = "1.6.7"
 
