### 2025-08-19 jake-w-liu/PlotlySupply.jl#3 @goerz
Clarify package name
We might as well get started on some of the preliminaries for the JOSS submission, even while we’re waiting for the second reviewer in https://github.com/openjournals/joss-reviews/issues/8471#issuecomment-3198418871…

Could you give some background on where the package name `PlotlySupply` came from? This is not to say that there is any issue with the name, but it might clarify the context of the package to understand what “Supply” is referring to here
