### 2025-08-20 jake-w-liu/PlotlySupply.jl#3 @jake-w-liu
Hello @goerz

Thank you for the question! The name **PlotlySupply** comes from the idea that this package was originally developed to *supply* my own research workflow with easier visualization tools on top of `PlotlyJS.jl`. While `PlotlyJS` provides comprehensive bindings, its API can be challenging for newcomers (at lease for me, like setting traces, layout, etc) and requires significant adjustments to produce figures suitable for research papers (e.g., white backgrounds, consistent axis settings, aspect ratios).

To streamline this process, I created **PlotlySupply.jl** with simple, MATLAB-like APIs that make it easier to generate common scientific plots directly, such as surface plots, scatter plots, heatmaps, and quiver plots (the latter not yet available in `PlotlyJS.jl`). In this sense, the package “supplies” the missing or simplified functionality needed to integrate interactive `Plotly` graphics into the research visualization flow.
