### 2025-08-26 #8471 @espottesmith
@chlalanne @altavir @vilim can any of y'all serve as reviewers?
