### 2025-08-26 jake-w-liu/PlotlySupply.jl#3 @goerz
Right… that's what I thought. I wanted to make sure it wasn't a reference to a technical term, like ["Supply" in economics](https://en.wikipedia.org/wiki/Supply_(economics)). I wouldn't say it's a great name. From a quasi-native-speaker perspective, it sounds a bit "weird". It might have been better to go with something like `HighLevelPlotly`, `PlotlyForHumans`, or something along those lines. But, especially since the package is registered in General already, it's probably not worth renaming at this point. What might be a good idea is to instead lean in to the name, and change the line in the README (and potentially also the project description in the "About" box) from

> **PlotlySupply.jl** is a lightweight Julia module that wraps PlotlyJS with high-level plotting APIs for 2D and 3D plot visualization.

to

> **PlotlySupply.jl** is a lightweight Julia module that wraps PlotlyJS to *supply* a set of high-level abstractions for 2D and 3D plot visualizations.
