### 2025-08-26 47a6466
Fix typos in JOSS paper
