### 2025-08-27 jake-w-liu/PlotlySupply.jl#3 @jake-w-liu
Thanks for pointing this out — I honestly hadn’t noticed the issue before 😊. I agree that “PlotlySupply” may sound a bit unusual, but since the package is already registered in General, I’m not sure renaming is possible? 🤔
I’ve updated the README accordingly, and the description now reads:

> PlotlySupply.jl is a lightweight Julia module that wraps PlotlyJS to supply a set of high-level abstractions for 2D and 3D plot visualizations.

Thanks again for the helpful suggestion!
