+++ README.md
@@ -5,7 +5,7 @@
 
 # PlotlySupply.jl
 
-**PlotlySupply.jl** is a lightweight Julia module that wraps PlotlyJS with high-level plotting APIs for 2D and 3D plot visualization. It simplifies the creation of line plots, heatmaps, quiver plots, polar plots, and surface plots, with consistent options for customization and layout control. Notice that this module re-exprots [`PlotlyJS.jl`](https://github.com/JuliaPlots/PlotlyJS.jl).
+**PlotlySupply.jl** is a lightweight Julia module that wraps PlotlyJS to supply a set of high-level abstractions for 2D and 3D plot visualizations. It simplifies the creation of line plots, heatmaps, quiver plots, polar plots, and surface plots, with consistent options for customization and layout control. Notice that this module re-exprots [`PlotlyJS.jl`](https://github.com/JuliaPlots/PlotlyJS.jl).
 
 ---
 
