+++ docs/make.jl
@@ -1,0 +1,5 @@
+using Documenter
+push!(LOAD_PATH,"../src/")
+using PlotlySupply
+
+makedocs(sitename="PlotlySupply.jl")
