+++ src/PlotlySupply.jl
@@ -7,9 +7,6 @@ using Infiltrator
 
 include("api.jl")
 
-# acronyms
-
-
 export plot_scatter, plot_stem, plot_scatterpolar, plot_heatmap, plot_quiver, plot_surface, plot_scatter3d, plot_quiver3d
 export set_template!
 
