+++ docs/src/index.md
@@ -1,0 +1,13 @@
+# PlotlySupply Documentation
+
+```@docs
+plot_scatter
+plot_stem
+plot_scatterpolar
+plot_heatmap
+plot_quiver
+plot_surface
+plot_scatter3d
+plot_quiver3d
+set_template!
+```
