+++ docs/make.jl
@@ -3,3 +3,7 @@ push!(LOAD_PATH,"../src/")
 using PlotlySupply
 
 makedocs(sitename="PlotlySupply.jl")
+
+deploydocs(
+    repo = "github.com/jake-w-liu/PlotlySupply.jl.git",
+)
