### 2025-08-27 22e3cee
add .woodpecker.yml
