+++ dev/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"#PlotlySupply-Documentation","page":"PlotlySupply Documentation","title":"PlotlySupply Documentation","text":"","category":"section"},{"location":"#PlotlySupply.plot_scatter","page":"PlotlySupply Documentation","title":"PlotlySupply.plot_scatter","text":"function plot_scatter( \tx::Union{AbstractRange, Vector}, \ty::Union{AbstractRange, Vector}; \txlabel::String = \"\", \tylabel::String = \"\", \txrange::Vector = [0, 0], \tyrange::Vector = [0, 0], \twidth::Int = 0, \theight::Int = 0, \tmode::Union{String, Vector{String}} = \"lines\", \tdash::Union{String, Vector{String}} = \"\", \tcolor::Union{String, Vector{String}} = \"\", \tlegend::Union{String, Vector{String}} = \"\", \ttitle::String = \"\", \tgrid::Bool = true, ) Plots a rectangular (Cartesian) plot.\n\nArguments\n\nx: x-coordinate data (can be vector of vectors)\ny: y-coordinate data (can be vector of vectors)\n\nKeywords\n\nxlabel: Label for the x-axis (default: \"\")\nylabel: Label for the y-axis (default: \"\")\nxrange: Range for the x-axis (default: [0, 0])\nyrange: Range for the y-axis (default: [0, 0])\nwidth: Width of the plot (default: 0)\nheight: Height of the plot (default: 0)\nmode: Plotting mode (default: \"lines\", can be vector)\ndash: line style (\"dash, \"dashdot\", or \"dot\", default: \"\", can be vector)\ncolor: Color of the plot lines (default: \"\", can be vector)\nlegend: Name of the plot lines (default: \"\", can be vector)\ntitle: Title of thje figure (default: \"\")\ngrid: Whether to show the grid or not (default: true)\n\n\n\n\n\nfunction plot_scatter( \ty::Union{AbstractRange, Vector};  \txlabel::String = \"\", \tylabel::String = \"\", \txrange::Vector = [0, 0], \tyrange::Vector = [0, 0], \twidth::Int = 0, \theight::Int = 0, \tmode::Union{String, Vector{String}} = \"lines\", \tdash::Union{String, Vector{String}} = \"\", \tcolor::Union{String, Vector{String}} = \"\", \tlegend::Union{String, Vector{String}} = \"\", \ttitle::String = \"\", \tgrid::Bool = true, ) Plots a rectangular (Cartesian) plot (x-axis not specified).\n\nArguments\n\ny: y-coordinate data (can be vector of vectors)\n\nKeywords\n\nxlabel: Label for the x-axis (default: \"\")\nylabel: Label for the y-axis (default: \"\")\nxrange: Range for the x-axis (default: [0, 0])\nyrange: Range for the y-axis (default: [0, 0])\nwidth: Width of the plot (default: 0)\nheight: Height of the plot (default: 0)\nmode: Plotting mode (default: \"lines\", can be vector)\ndash: line style (\"dash, \"dashdot\", or \"dot\", default: \"\", can be vector)\ncolor: Color of the plot lines (default: \"\", can be vector)\nlegend: legend of the plot lines (default: \"\", can be vector)\ntitle: Title of thje figure (default: \"\")\ngrid: Whether to show the grid or not (default: true)\n\n\n\n\n\n","category":"function"},{"location":"#PlotlySupply.plot_stem","page":"PlotlySupply Documentation","title":"PlotlySupply.plot_stem","text":"function plot_stem( \tx::Union{AbstractRange, Vector}, \ty::Union{AbstractRange, Ve
[…]
