+++ docs/make.jl
@@ -1,8 +1,12 @@
+using Pkg
+Pkg.activate("..")
+
 using Documenter
-push!(LOAD_PATH,"../src/")
 using PlotlySupply
 
-makedocs(sitename="PlotlySupply.jl")
+makedocs(
+    sitename="PlotlySupply.jl",
+)
 
 deploydocs(
     repo = "github.com/jake-w-liu/PlotlySupply.jl.git",
