+++ docs/src/index.md
@@ -1,13 +1,6 @@
 # PlotlySupply Documentation
 
-```@docs
-plot_scatter
-plot_stem
-plot_scatterpolar
-plot_heatmap
-plot_quiver
-plot_surface
-plot_scatter3d
-plot_quiver3d
-set_template!
+```@autodocs
+Modules = [PlotlySupply]
+Order   = [:function, :type]
 ```
