+++ docs/make.jl
@@ -1,5 +1,3 @@
-using Pkg
-Pkg.activate("..")
 
 using Documenter
 using PlotlySupply
