+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"#PlotlySupply-Documentation","page":"PlotlySupply Documentation","title":"PlotlySupply Documentation","text":"","category":"section"},{"location":"#PlotlySupply.plot_heatmap-Tuple{Array}","page":"PlotlySupply Documentation","title":"PlotlySupply.plot_heatmap","text":"function plot_heatmap( \tU::Array;  \txlabel::String = \"\", \tylabel::String = \"\", \txrange::Vector = [0, 0], \tyrange::Vector = [0, 0], \tzrange::Vector = [0, 0], \twidth::Int = 0, \theight::Int = 0, \tcolorscale::String = \"Jet\", \ttitle::String = \"\", \tequalar::Bool = false, ) Plots holographic data.\n\nArguments\n\nU: 2D hologram data\n\nKeywords\n\nxlabel: Label for the x-axis (default: \"\")\nylabel: Label for the y-axis (default: \"\")\nzrange: Range for the z-axis (default: [0, 0])\ncolorscale: Color scale for the heatmap (default: \"Jet\")\ntitle: Title of thje figure (default: \"\")\n\n\n\n\n\n","category":"method"},{"location":"#PlotlySupply.plot_heatmap-Tuple{Union{AbstractRange, Vector}, Union{AbstractRange, Vector}, Array}","page":"PlotlySupply Documentation","title":"PlotlySupply.plot_heatmap","text":"function plot_heatmap( \tx::Union{AbstractRange, Vector}, \ty::Union{AbstractRange, Vector}, \tU::Array;  \txlabel::String = \"\", \tylabel::String = \"\", \txrange::Vector = [0, 0], \tyrange::Vector = [0, 0], \tzrange::Vector = [0, 0], \twidth::Int = 0, \theight::Int = 0, \tcolorscale::String = \"Jet\", \ttitle::String = \"\", \tequalar::Bool = false, )\n\nPlots holographic data.\n\nArguments\n\n'x::AbstractRange': x-axis range\n'y::AbstractRange': x-axis range\nU: 2D hologram data\n\nKeywords\n\nxlabel: Label for the x-axis (default: \"\")\nylabel: Label for the y-axis (default: \"\")\nzrange: Range for the z-axis (default: [0, 0])\ncolorscale: Color scale for the heatmap (default: \"Jet\")\ntitle: Title of thje figure (default: \"\")\n\n\n\n\n\n","category":"method"},{"location":"#PlotlySupply.plot_quiver-NTuple{4, Union{AbstractRange, Vector}}","page":"PlotlySupply Documentation","title":"PlotlySupply.plot_quiver","text":"function plot_quiver(     x::Union{AbstractRange, Vector},     y::Union{AbstractRange, Vector},     u::Union{AbstractRange, Vector},     v::Union{AbstractRange, Vector};     color::String = \"RoyalBlue\",     sizeref::Real = 1,     xlabel::String = \"\",     ylabel::String = \"\",     xrange::Vector = [0, 0],     yrange::Vector = [0, 0],     width::Int = 0,     height::Int = 0,     title::String = \"\",     grid::Bool = true, ) Plots a 2D quiver (vector field) diagram using arrow segments.\n\nArguments\n\nx: x-coordinates of vector origins\ny: y-coordinates of vector origins\nu: x-components of vector directions\nv: y-components of vector directions\n\nKeywords\n\ncolor: Arrow color (default: \"RoyalBlue\")\nsizeref: Reference scaling for arrow length (default: 1)\nxlabel: Label for the x-axis (default: \"\")\nylabel: Label for the y-axis (default: \"\")\nxrange: Range for
[…]
