+++ src/api.jl
@@ -17,7 +17,7 @@
 		title::String = "",
 		grid::Bool = true,
 	)
-		
+
 Plots a rectangular (Cartesian) plot.
 
 #### Arguments
@@ -160,21 +160,22 @@ function plot_scatter(
 end
 
 """
-function plot_scatter(
-	y::Union{AbstractRange, Vector}; 
-	xlabel::String = "",
-	ylabel::String = "",
-	xrange::Vector = [0, 0],
-	yrange::Vector = [0, 0],
-	width::Int = 0,
-	height::Int = 0,
-	mode::Union{String, Vector{String}} = "lines",
-	dash::Union{String, Vector{String}} = "",
-	color::Union{String, Vector{String}} = "",
-	legend::Union{String, Vector{String}} = "",
-	title::String = "",
-	grid::Bool = true,
-)
+	function plot_scatter(
+		y::Union{AbstractRange, Vector}; 
+		xlabel::String = "",
+		ylabel::String = "",
+		xrange::Vector = [0, 0],
+		yrange::Vector = [0, 0],
+		width::Int = 0,
+		height::Int = 0,
+		mode::Union{String, Vector{String}} = "lines",
+		dash::Union{String, Vector{String}} = "",
+		color::Union{String, Vector{String}} = "",
+		legend::Union{String, Vector{String}} = "",
+		title::String = "",
+		grid::Bool = true,
+	)
+
 Plots a rectangular (Cartesian) plot (x-axis not specified).
 
 #### Arguments
@@ -240,20 +241,22 @@ function plot_scatter(
 end
 
 """
-function plot_stem(
-	x::Union{AbstractRange, Vector},
-	y::Union{AbstractRange, Vector};
-	xlabel::String = "",
-	ylabel::String = "",
-	xrange::Vector = [0, 0],
-	yrange::Vector = [0, 0],
-	width::Int = 0,
-	height::Int = 0,
-	color::Union{String, Vector{String}} = "",
-	legend::Union{String, Vector{String}} = "",
-	title::String = "",
-	grid::Bool = true,
-)
+	function plot_stem(
+		x::Union{AbstractRange, Vector},
+		y::Union{AbstractRange, Vector};
+		xlabel::String = "",
+		ylabel::String = "",
+		xrange::Vector = [0, 0],
+		yrange::Vector = [0, 0],
+		width::Int = 0,
+		height::Int = 0,
+		color::Union{String, Vector{String}} = "",
+		legend::Union{String, Vector{String}} = "",
+		title::String = "",
+		grid::Bool = true,
+	)
+
+Plots a rectangular stem plot.
 
 #### Arguments
 
@@ -453,19 +456,21 @@ function plot_stem(
 end
 
 """
-function plot_stem(
-	y::Union{AbstractRange, Vector};
-	xlabel::String = "",
-	ylabel::String = "",
-	xrange::Vector = [0, 0],
-	yrange::Vector = [0, 0],
-	width::Int = 0,
-	height::Int = 0,
-	color::Union{String, Vector{String}} = "",
-	legend::Union{String, Vector{String}} = "",
-	title::String = "",
-	grid::Bool = true,
-)
+	function plot_stem(
+		y::Union{AbstractRange, Vector};
+		xlabel::String = "",
+		ylabel::String = "",
+		xrange::Vector = [0, 0],
+		yrange::Vector = [0, 0],
+		width::Int = 0,
+		height::Int = 0,
+		color::Union{String, Vector{String}} = "",
+		legend::Union{String, Vector{String}} = "",
+		title::String = "",
+		grid::Bool = true,
+	)
+
+Plots a rectangular stem plot (x-axis not specified).
 
 #### Arguments
 
[…]
