+++ .github/workflows/CI.yml
@@ -39,3 +39,22 @@ jobs:
       - uses: julia-actions/cache@v2
       - uses: julia-actions/julia-buildpkg@v1
       - uses: julia-actions/julia-runtest@v1
+      - uses: julia-actions/julia-processcoverage@v1
+      - name: Install LocalCoverage
+        shell: julia {0}
+        run: |
+          using Pkg
+          Pkg.add(PackageSpec(name="LocalCoverage"))
+          Pkg.add(PackageSpec(name="CoverageTools"))
+      - name: Show coverage info
+        shell: julia {0}
+        run: |
+          import CoverageTools
+          import LocalCoverage
+          coverage = CoverageTools.LCOV.readfile("lcov.info")
+          report = LocalCoverage.eval_coverage_metrics(coverage, ".")
+          show(report)
+      - uses: codecov/codecov-action@v5
+        with:
+          files: lcov.info
+          token: ${{ secrets.CODECOV_TOKEN }}
