+++ test/runtests.jl
@@ -1,56 +1,122 @@
-using PlotlySupply
 using Test
-using PlotlyJS
+using PlotlySupply
 
 @testset "PlotlySupply.jl" begin
+    # Write your tests here.
+
     @testset "plot_scatter" begin
-        @test plot_scatter(1:10, 1:10) isa PlotlyJS.SyncPlot
-        @test plot_scatter(1:10) isa PlotlyJS.SyncPlot
-        @test plot_scatter(1:10, 1:10, xlabel="x", ylabel="y", title="title", legend="legend") isa PlotlyJS.SyncPlot
-        @test plot_scatter(1:10, [1:10, 2:11], mode=["lines", "markers"]) isa PlotlyJS.SyncPlot
-    end
+        x = 1:10
+        y = rand(10)
+        fig = plot_scatter(x, y, title="My Title", xlabel="x", ylabel="y", xrange=[0, 10], yrange=[0, 1], width=500, height=500, grid=false)
+        @test fig isa PlotlyJS.SyncPlot
 
-    @testset "plot_heatmap" begin
-        @test plot_heatmap(rand(10, 10)) isa PlotlyJS.SyncPlot
-        @test plot_heatmap(1:10, 1:10, rand(10, 10)) isa PlotlyJS.SyncPlot
-        @test plot_heatmap(1:10, 1:10, rand(10, 10), xlabel="x", ylabel="y", title="title") isa PlotlyJS.SyncPlot
-    end
-
-    @testset "plot_surface" begin
-        @test plot_surface(rand(10,10)) isa PlotlyJS.SyncPlot
-        @test plot_surface(collect(1:10), collect(1:10), rand(10,10)) isa PlotlyJS.SyncPlot
-        @test plot_surface(collect(1:10), collect(1:10), rand(10,10), xlabel="x", ylabel="y", zlabel="z", title="title") isa PlotlyJS.SyncPlot
-    end
-
-    @testset "plot_quiver" begin
-        @test plot_quiver(1:10, 1:10, 1:10, 1:10) isa PlotlyJS.SyncPlot
-        @test plot_quiver(1:10, 1:10, 1:10, 1:10, xlabel="x", ylabel="y", title="title") isa PlotlyJS.SyncPlot
-    end
-
-    @testset "plot_scatter3d" begin
-        @test plot_scatter3d(1:10, 1:10, 1:10) isa PlotlyJS.SyncPlot
-        @test plot_scatter3d(1:10, 1:10, 1:10, xlabel="x", ylabel="y", zlabel="z", title="title") isa PlotlyJS.SyncPlot
-    end
-
-    @testset "plot_quiver3d" begin
-        @test plot_quiver3d(1:10, 1:10, 1:10, 1:10, 1:10, 1:10) isa PlotlyJS.SyncPlot
-        @test plot_quiver3d(1:10, 1:10, 1:10, 1:10, 1:10, 1:10, xlabel="x", ylabel="y", zlabel="z", title="title") isa PlotlyJS.SyncPlot
+        y2 = [rand(10), rand(10)]
+        fig2 = plot_scatter(x, y2)
+        @test fig2 isa PlotlyJS.SyncPlot
     end
 
     @testset "plot_stem" begin
-        @test plot_stem(1:10, 1:10) isa PlotlyJS.SyncPlot
-        @test plot_stem(1:10) isa PlotlyJS.SyncPlot
-        @test plot_stem(1:10, 1:10, xlabel="x", ylabel="y", title="title", legend="legend") isa PlotlyJS.SyncPlot
+        x = 1:10
+        y = rand(10)
[…]
