### 2025-08-29 3b612dc
update readme (add documentation link)
