### 2025-08-30 6150d61
try to add interactive example
