### 2025-08-30 1ab9386
add example figures
