### 2025-08-30 a0fcbd7
add example figures
