### 2025-08-30 168e0d4
change example fig urls
