### 2025-08-30 714e8ec
update cleanup in doc yml
