+++ .github/workflows/documentation.yml
@@ -3,15 +3,25 @@ name: Documentation
 on:
   push:
     branches:
-      - master # update to match your development branch (master, main, dev, trunk, ...)
-    tags: '*'
+      - main # update to match your development branch (master, main, dev, trunk, ...)
   pull_request:
   workflow_dispatch:
 jobs:
+  cleanup:
+      runs-on: ubuntu-latest
+      permissions: write-all
+      steps:
+        - name: Delete deployment
+          uses: strumwolf/delete-deployment-environment@v2
+          with:
+            token: ${{ secrets.GITHUB_TOKEN }}
+            environment: github-pages
+            onlyRemoveDeployments: true
   build:
     # These permissions are needed to:
     # - Deploy the documentation: https://documenter.juliadocs.org/stable/man/hosting/#Permissions
     # - Delete old caches: https://github.com/julia-actions/cache#usage
+    needs: cleanup
     permissions:
       actions: write
       contents: write
