### 2025-08-31 22dd0e0
add figs to example
