+++ docs/src/example.md
@@ -1,6 +1,14 @@
-# Examples
+# Plotting Examples with PlotlySupply.jl
+
+This page provides a gallery of examples to showcase the plotting capabilities of `PlotlySupply.jl`. Each example includes the Julia code used to generate the plot and the resulting figure. These examples are designed to help you get started with the package and demonstrate how to create a variety of common plot types.
 
 ## `plot_scatter`
+
+Scatter plots are versatile and can be used to visualize relationships between variables, such as in line charts, scatter plots, or a combination of both.
+
+### Simple Line Plot
+
+This example demonstrates how to create a simple line plot of a quadratic function. We define our x and y coordinates and then pass them to `plot_scatter` with labels for the axes and a title.
 
 ```julia
 # Simple line plot
@@ -11,12 +19,17 @@ fig = plot_scatter(x, y, xlabel="X", ylabel="Y²", title="Quadratic Function")
 
 [![ScatterPlot](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter.png)](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter.html)
 
+### Multiple Lines with Different Styles
+
+Here, we showcase how to plot multiple lines on the same axes. We can customize the appearance of each line by specifying the `mode`, `dash`, `color`, and `legend` for each dataset.
+
 ```julia
 # Multiple lines with different styles
+x = 1:10
 y1 = sin.(x)
 y2 = cos.(x)
 fig = plot_scatter(x, [y1, y2], 
-                   mode=["lines", "lines"], 
+                   mode=["lines+markers", "lines"], 
                    dash=["", "dash"],
                    color=["blue", "red"],
                    legend=["sin(x)", "cos(x)"])
@@ -25,6 +38,12 @@ fig = plot_scatter(x, [y1, y2],
 [![ScatterPlot2](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter2.png)](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter2.html)
 
 ## `plot_stem`
+
+Stem plots are ideal for visualizing discrete signals, showing data points as stems extending from a baseline.
+
+### Discrete Signal Visualization
+
+This example illustrates how to create a stem plot for a simple discrete signal. This is particularly useful in signal processing to visualize the amplitude of a signal at different sample points.
 
 ```julia
 # Discrete signal visualization
@@ -37,6 +56,12 @@ fig = plot_stem(n, signal, xlabel="Sample", ylabel="Amplitude")
 
 ## `plot_scatterpolar`
 
+Polar plots are used to visualize data in a polar coordinate system, which is useful for cyclical data or directional data.
+
+### Polar Rose Pattern
+
+This example generates a beautiful rose pattern using a polar plot. The angle `theta` and radius `r` are used to create this intricate shape.
+
 ```julia
 # Polar rose pattern
 theta = 0:1:360
@@ -47,6 +72,12 @@ fig = plot_scatterpolar(theta, r, title="Rose Pattern")
 [![ScatterpolarPlot](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatterpolar.png)](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatterpolar.html)
 
 ## `plot_heatmap`
+
+Heatmaps are excellent for visualizing 2D data, where values are represented by colors.
+
+### 2D Gaussian Distribution
+
[…]
