+++ docs/src/api.md
@@ -1,0 +1,6 @@
+# API Documentation
+
+```@autodocs
+Modules = [PlotlySupply]
+Order   = [:function, :type]
+```
