+++ docs/src/example.md
@@ -17,7 +17,7 @@ y = x.^2
 fig = plot_scatter(x, y, xlabel="X", ylabel="Y²", title="Quadratic Function")
 ```
 
-[![ScatterPlot](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter.png)](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter.html)
+<a href="https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter.html" target="_blank"><img src="https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter.png" alt="ScatterPlot" style="width:60%; display: block; margin: auto;"></a>
 
 ### Multiple Lines with Different Styles
 
@@ -35,7 +35,7 @@ fig = plot_scatter(x, [y1, y2],
                    legend=["sin(x)", "cos(x)"])
 ```
 
-[![ScatterPlot2](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter2.png)](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter2.html)
+<a href="https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter2.html" target="_blank"><img src="https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter2.png" alt="ScatterPlot2" style="width:60%; display: block; margin: auto;"></a>
 
 ## `plot_stem`
 
@@ -52,7 +52,7 @@ signal = [1, 0, -1, 0, 1, 0, -1, 0, 1, 0, -1]
 fig = plot_stem(n, signal, xlabel="Sample", ylabel="Amplitude")
 ```
 
-[![StemPlot](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_stem.png)](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_stem.html)
+<a href="https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_stem.html" target="_blank"><img src="https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_stem.png" alt="StemPlot" style="width:60%; display: block; margin: auto;"></a>
 
 ## `plot_scatterpolar`
 
@@ -69,7 +69,7 @@ r = 3 .* sind.(4 * theta)
 fig = plot_scatterpolar(theta, r, title="Rose Pattern")
 ```
 
-[![ScatterpolarPlot](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatterpolar.png)](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatterpolar.html)
+<a href="https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatterpolar.html" target="_blank"><img src="https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatterpolar.png" alt="ScatterpolarPlot" style="width:60%; display: block; margin: auto;"></a>
 
 ## `plot_heatmap`
 
@@ -88,7 +88,7 @@ Z = exp.(-(X.^2 + Y.^2))
 fig = plot_heatmap(x, y, Z, title="2D Gaussian", colorscale="Viridis", equalar=true)
 ```
 
-[![HeatmapPlot](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_heatmap.png)](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_heatmap.html)
[…]

[…]
