+++ docs/src/example.md
@@ -1,4 +1,4 @@
-# Plotting Examples with PlotlySupply.jl
+# Examples 
 
 This page provides a gallery of examples to showcase the plotting capabilities of `PlotlySupply.jl`. Each example includes the Julia code used to generate the plot and the resulting figure. These examples are designed to help you get started with the package and demonstrate how to create a variety of common plot types.
 
@@ -17,7 +17,7 @@ y = x.^2
 fig = plot_scatter(x, y, xlabel="X", ylabel="Y²", title="Quadratic Function")
 ```
 
-<a href="https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter.html" target="_blank"><img src="https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter.png" alt="ScatterPlot" style="width:60%; display: block; margin: auto;"></a>
+[![ScatterPlot](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter.png)](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter.html)
 
 ### Multiple Lines with Different Styles
 
@@ -35,7 +35,7 @@ fig = plot_scatter(x, [y1, y2],
                    legend=["sin(x)", "cos(x)"])
 ```
 
-<a href="https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter2.html" target="_blank"><img src="https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter2.png" alt="ScatterPlot2" style="width:60%; display: block; margin: auto;"></a>
+[![ScatterPlot2](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter2.png)](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatter2.html)
 
 ## `plot_stem`
 
@@ -52,7 +52,7 @@ signal = [1, 0, -1, 0, 1, 0, -1, 0, 1, 0, -1]
 fig = plot_stem(n, signal, xlabel="Sample", ylabel="Amplitude")
 ```
 
-<a href="https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_stem.html" target="_blank"><img src="https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_stem.png" alt="StemPlot" style="width:60%; display: block; margin: auto;"></a>
+[![StemPlot](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_stem.png)](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_stem.html)
 
 ## `plot_scatterpolar`
 
@@ -69,7 +69,7 @@ r = 3 .* sind.(4 * theta)
 fig = plot_scatterpolar(theta, r, title="Rose Pattern")
 ```
 
-<a href="https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatterpolar.html" target="_blank"><img src="https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatterpolar.png" alt="ScatterpolarPlot" style="width:60%; display: block; margin: auto;"></a>
+[![ScatterpolarPlot](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatterpolar.png)](https://jake-w-liu.github.io/assets/img/PlotlySupply/fig_scatterpolar.html)
 
 ## `plot_heatmap`
 
@@ -88,7 +88,7 @@ Z = exp.(-(X.^2 + Y.^2))
[…]
