+++ Project.toml
@@ -1,7 +1,7 @@
 name = "PlotlySupply"
 uuid = "a389b534-8bca-4fe3-8ff2-14829cd8824f"
 authors = ["Jake W. Liu"]
-version = "1.1.0"
+version = "1.1.1"
 
 [deps]
 BatchAssign = "e837618a-1c15-48db-beb0-68297dffdb28"
