+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"api/#API-Documentation","page":"API Documentation","title":"API Documentation","text":"","category":"section"},{"location":"api/#PlotlySupply.plot_heatmap-Tuple{Array}","page":"API Documentation","title":"PlotlySupply.plot_heatmap","text":"function plot_heatmap(\n\tU::Array; \n\txlabel::String = \"\",\n\tylabel::String = \"\",\n\txrange::Vector = [0, 0],\n\tyrange::Vector = [0, 0],\n\tzrange::Vector = [0, 0],\n\twidth::Int = 0,\n\theight::Int = 0,\n\tcolorscale::String = \"Jet\",\n\ttitle::String = \"\",\n\tequalar::Bool = false,\n)\n\nPlots heatmap (holographic) data (axes not specified).\n\nArguments\n\nU: 2D hologram data\n\nKeywords\n\nxlabel: Label for the x-axis (default: \"\")\nylabel: Label for the y-axis (default: \"\")\nzrange: Range for the z-axis (default: [0, 0])\ncolorscale: Color scale for the heatmap (default: \"Jet\")\ntitle: Title of thje figure (default: \"\")\n\n\n\n\n\n","category":"method"},{"location":"api/#PlotlySupply.plot_heatmap-Tuple{Union{AbstractRange, Vector}, Union{AbstractRange, Vector}, Array}","page":"API Documentation","title":"PlotlySupply.plot_heatmap","text":"function plot_heatmap(\n\tx::Union{AbstractRange, Vector},\n\ty::Union{AbstractRange, Vector},\n\tU::Array; \n\txlabel::String = \"\",\n\tylabel::String = \"\",\n\txrange::Vector = [0, 0],\n\tyrange::Vector = [0, 0],\n\tzrange::Vector = [0, 0],\n\twidth::Int = 0,\n\theight::Int = 0,\n\tcolorscale::String = \"Jet\",\n\ttitle::String = \"\",\n\tequalar::Bool = false,\n)\n\nPlots heatmap (holographic) data.\n\nArguments\n\n'x::AbstractRange': x-axis range\n'y::AbstractRange': x-axis range\nU: 2D hologram data\n\nKeywords\n\nxlabel: Label for the x-axis (default: \"\")\nylabel: Label for the y-axis (default: \"\")\nzrange: Range for the z-axis (default: [0, 0])\ncolorscale: Color scale for the heatmap (default: \"Jet\")\ntitle: Title of thje figure (default: \"\")\n\n\n\n\n\n","category":"method"},{"location":"api/#PlotlySupply.plot_quiver-NTuple{4, Union{AbstractRange, Vector}}","page":"API Documentation","title":"PlotlySupply.plot_quiver","text":"function plot_quiver(\n\tx::Union{AbstractRange, Vector},\n\ty::Union{AbstractRange, Vector},\n\tu::Union{AbstractRange, Vector},\n\tv::Union{AbstractRange, Vector};\n\tcolor::String = \"RoyalBlue\",\n\tsizeref::Real = 1,\n\txlabel::String = \"\",\n\tylabel::String = \"\",\n\txrange::Vector = [0, 0],\n\tyrange::Vector = [0, 0],\n\twidth::Int = 0,\n\theight::Int = 0,\n\ttitle::String = \"\",\n\tgrid::Bool = true,\n)\n\nPlots a 2D quiver (vector field) diagram using arrow segments.\n\nArguments\n\nx: x-coordinates of vector origins\ny: y-coordinates of vector origins\nu: x-components of vector directions\nv: y-components of vector directions\n\nKeywords\n\ncolor: Arrow color (defau
[…]
