### 2025-09-01 #8902 @editorialbot
[REVIEW]: PlotlySupply.jl: High-level research visualization wrappers in Julia with a user-friendly API, built on PlotlyJS.jl
**Submitting author:** @jake-w-liu (<a href="http://orcid.org/0000-0001-5458-7917">Jake W. Liu</a>)
**Repository:** https://github.com/jake-w-liu/PlotlySupply.jl
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
