### 2025-09-01 #8902 @goerz
> the software is sufficiently useful that it is likely to be cited by your peer group

I kinda missed that last bullet point, but from that perspective, I would agree. There is basically no circumstance where I would cite a general plotting package as part of an academic paper. It just wouldn't be related to the scientific story presented in a paper (assuming it's not a CS paper about visualization). That would apply to the original PlotlyJS just as much as to any wrapper around it. I could still see wanting to publish a JOSS paper from the perspective of "I spent a bunch of time on this, and I want a publication record that I can include in a funding review"

If I was an editor, I would have voted against this package being in scope for JOSS, but I took [the result of the scope review](https://github.com/openjournals/joss-reviews/issues/8471#issuecomment-3180248434) as a given, and wasn't going to further comment on it. I'd again defer to the editor's judgement on this.
