### 2025-09-01 #8902 @jolars
Thanks, but even so, this still seems like something that should be voiced in the review as well:

> Reviewers should verify that the software represents substantial scholarly effort.

I will of course defer judgment to the editors on this, but, hm, I cannot help but feel that this doesn't quite pass the test in terms of "substantial scholarly effort".

> Your software should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler).

I'm not entirely sure the examples demonstrate this. For instance, going from

```julia
plot_surface(X, Y, Z; title = "3D Surface", colorscale = "Viridis")
```

to

```julia
plot(surface(x = X, y = Y, z = Z, colorscale = "Viridis"), Layout(title = "3D Surface"))
```

is not, at least not to me, a significant improvement. In fact once could easily make the opposite argument.
