### 2025-09-02 #8902 @SunnyXu
@espottesmith @jake-w-liu
PlotlySupply.jl provided a visualization wrappers in Julia built on PlotlyJS.jl.
-Because this work is highly based on PlotlyJS.jl, I think it is meaningful to hightlight its advantage beyond PlotlyJS.jl with examples in the paper, readme, and/or the introduction part of documentation.
-The author mentioned that "the API design of PlotlySupply.jl is intentionally similar to MATLAB-style plotting, making it intuitive for researchers transitioning to Julia or working in multi-language environments." It maybe deserved to discuss about this further too.
-For the Research Applications, the author self-cited two papers from himself. Are there potential applications by others?
Maybe these discussions above could make it more eligible to get published in JOSS.

Please also see the unchecked items in my checklist for possible improvement. Thanks.
