+++ CONTRIBUTING.md
@@ -1,0 +1,49 @@
+# Contributing to PlotlySupply.jl
+
+First off, thank you for considering contributing to PlotlySupply.jl. It's people like you that make open source such a great community.
+
+## Where do I go from here?
+
+If you've noticed a bug or have a feature request, [make one](https://github.com/jake-w-liu/PlotlySupply.jl/issues/new)! It's generally best if you get confirmation of your bug or approval for your feature request this way before starting to code.
+
+### Reporting Bugs
+
+When reporting a bug, please include as much detail as possible. This might include:
+
+- A description of the bug and the behavior you expected.
+- A minimal, reproducible example.
+- The version of PlotlySupply.jl you are using.
+- Your Julia version.
+
+### Suggesting Enhancements
+
+If you have an idea for a new feature, please provide a clear and detailed explanation of the proposed enhancement. Include why you think it would be a valuable addition to the project and, if possible, provide a code snippet or pseudo-code demonstrating its usage.
+
+## Project Scope
+
+PlotlySupply.jl aims to be a lightweight wrapper around PlotlyJS.jl, providing high-level abstractions for common 2D and 3D plots. The goal is to simplify the creation of plots with a consistent and intuitive API.
+
+Contributions should align with this scope. We are particularly interested in:
+
+- Improving existing plotting functions.
+- Adding new high-level plotting functions that are commonly used.
+- Enhancing documentation and examples.
+- Fixing bugs and improving performance.
+
+We are generally not looking to:
+
+- Add highly specialized or niche plot types.
+- Replicate the full functionality of PlotlyJS.jl.
+- Introduce significant new dependencies.
+
+## Development
+
+1.  Fork the repository and create a new branch from `main`.
+2.  Make your changes.
+3.  Add tests for your changes.
+4.  Ensure your code is well-documented.
+5.  Open a pull request.
+
+## License
+
+By contributing, you agree that your contributions will be licensed under the MIT License.
