+++ paper/paper.bib
@@ -1,8 +1,10 @@
-@article{sievert2021package,
-  title={Package ‘plotly’},
-  author={Sievert, Carson and Parmer, Chris and Hocking, Toby and Chamberlain, Scott and Ram, Karthik and Corvellec, Marianne and Despouy, Pedro and Sievert, Maintainer Carson},
-  journal={R Foundation for Statistical Computing, Vienna. Available in: https://cran. r-project. org/package= plotly (accessed 10 August 2023)},
-  year={2021}
+@online{plotly, 
+  author = {Plotly Technologies Inc.}, 
+  title = {Collaborative data science}, 
+  publisher = {Plotly Technologies Inc.}, 
+  address = {Montreal, QC}, 
+  year = {2015}, 
+  url = {https://plot.ly} 
 }
 
 @article{liu2024near,
