### 2025-09-02 014f1d9
update paper - add comparison
