### 2025-09-02 #8902 @espottesmith
This code/manuscript was originally flagged by the editorial board; there was a concern about scholarly effort. As noted by @goerz above, the editorial board held a vote, and ultimately it was decided that this did pass the bar.

Our original determination is not the final word. @jolars is correct that scholarly effort can and should also be addressed by reviewers. After all, we put that line in the "General checks" category of the reviewer checklist for a reason.

Notably, all three reviewers have indicated that, in the current form, `PlotlySupply.jl` does not meet standards of substantial scholarly effort. As an editor, I am generally inclined to support any unanimous recommendation from the assigned reviewers. And I think that the apparent consensus in this case, coupled with the original uncertainty raised by the editorial board, makes it clear that, in its present form, this package should not be considered for publication in JOSS.

I'll flag the Associate Editor-in-Chief and recommend rejection.
