### 2025-09-02 #8902 @kyleniemeyer
Hi all - I agree with the discussion and consensus here. Thanks @goerz, @SunnyXu, and @jolars for your efforts and feedback here, and for @espottesmith for managing this.

@jake-w-liu though it sounds like you have some plans to continue working on your software package, in the meantime I encourage you to pursue another venue for citing/receiving credit for your work:
- Follow [GitHub’s guide](https://guides.github.com/activities/citable-code/) on how to create a permanent archive and DOI for your software. This DOI can then be used by others to cite your work.
- Enquire whether your software might be considered by a community such as [pyOpenSci](https://pyopensci.org/).
